Redo the setup.py
LICENSE is now included as a separated file, in the
project's root directory. Setup now takes cares of 
distributing the license with the installation and
its also added in the PKG-INFO file in the xmsg python
egg.
Version is now defined in the xmsg.__init__.py file
in the __version__ variable.